requirements.txt
flask
transformers
torch
protobuf 
sentencepiece
openai
python-dotenv
datasets